requirements.txt
absl-py==1.3.0
anyio==3.6.2
APScheduler==3.6.3
args==0.1.0
astroid==2.4.2
astunparse==1.6.3
attrs==22.1.0
autopep8==1.5.4
backcall==0.2.0
backoff==1.10.0
backports.csv==1.0.7
backports.shutil-get-terminal-size==1.0.0
basis-expansions==0.0.1
biopython==1.78
CacheControl==0.12.11
cachetools==4.2.2
certifi==2017.7.27.1
cffi==1.15.1
chardet==3.0.4
click==6.7
clint==0.5.1
colorama==0.4.3
configparser==3.5.0
contextlib2==21.6.0
cryptography==38.0.1
cssselect2==0.3.0
cycler==0.10.0
debugpy==1.6.3
decorator==4.4.2
docopt==0.6.2
entrypoints==0.4
enum34==1.1.6
et-xmlfile==1.1.0
fastaparser==1.1
firebase-admin==6.0.1
flake8==3.4.1
Flask==1.0.2
flatbuffers==22.12.6
gast==0.4.0
geojson==2.5.0
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.50.0
gunicorn==19.9.0
h11==0.12.0
h5py==3.7.0
httpcore==0.15.0
httplib2==0.21.0
httpx==0.23.0
idna==2.10
imageio==2.22.4
importlib-metadata==5.0.0
internetarchive==1.9.0
ipykernel==6.16.0
ipython==7.34.0
isort==5.4.2
itsdangerous==0.24
jedi==0.18.1
Jinja2==2.10
joblib==1.2.0
jsonpatch==1.32
jsonpickle==0.9.3
jsonpointer==2.3
jsonschema==3.2.0
jupyter-core==4.11.1
jupyter_client==7.4.2
keras==2.11.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
libclang==14.0.6
lxml==4.5.2
Markdown==2.6.11
MarkupSafe==2.1.1
matplotlib==3.4.0
matplotlib-inline==0.1.6
mccabe==0.6.1
mlxtend==0.21.0
msgpack==1.0.4
nest-asyncio==1.5.6
networkx==2.5
nose==1.3.7
numpy==1.20.2
oauthlib==3.2.2
opencv-python==4.5.1.48
openlibrary-client==0.0.30
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.2.3
parso==0.8.3
pathlib==1.0.1
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.31
proto-plus==1.22.1
protobuf==3.9.2
psutil==5.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.3.1
pycparser==2.21
pyflakes==1.5.0
Pygments==2.13.0
PyJWT==2.6.0
pylast==5.1.0
pylint==2.6.0
pymarc==3.2.0
pyOpenSSL==22.1.0
pyowm==3.3.0
pyparsing==2.4.7
PyQt5==5.13.1
pyqt5-tools==5.13.1.1.6rc0
PyQt5_sip==4.19.19
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.10.3
python-telegram-bot==13.7
pytz==2021.1
PyWavelets==1.3.0
pywin32==304
pyzmq==24.0.1
reportlab==3.5.51
requests==2.23.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
schema==0.7.5
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
statsmodels==0.13.2
svglib==1.0.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.0.2
toml==0.10.1
torch==1.8.1
tornado==6.2
traitlets==5.4.0
typed-ast==1.4.1
typing-extensions==3.10.0.0
tzlocal==2.1
uritemplate==4.1.1
urllib3==1.25.11
virtualenv-clone==0.2.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.12.1
xlrd==2.0.1
zipp==3.10.0
